feat : util 함수 따로 빼기
